Rename ImageGallery to ImageCarousel Make ImageGallery display 3 images side by side
