+++ src/lib/imageGallery.ts
@@ -1,4 +1,8 @@
 export type Image = {
         url: string,
         description: string,
-    }
+}
+
+export function randomInteger(min: number, max: number) {
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+  }
